# backend/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "APP2000 API",
		"version": "1.0.0",
		"description": "Full API-dokumentasjon for brukere, klubber, turneringer, systemlogg m.m."
	},
	"servers": [
		{ "url": "/" }
	],
	"paths": {
			"/sjekk-session": {
				"get": {
					"summary": "Sjekk om session er aktiv og hent brukerdata",
					"security": [{ "cookieAuth": [] }],
					"responses": {
						"200": { "description": "Innlogget", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SessionBruker" } } } },
						"401": { "description": "Ingen aktiv session" },
						"500": { "description": "Serverfeil" }
					}
				}
			},
...
		"/api/systemlogg": {
			"get": { "summary": "Hent systemlogg", "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Systemlogg" } } } } } } },
			"post": { "summary": "Logg hendelse", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SystemloggInput" } } } }, "responses": { "201": { "description": "Opprettet" }, "400": { "description": "Mangler felt" } } }
		},
...
			"/KontaktOss": {
				"post": {
					"summary": "Send kontaktskjema",
					"requestBody": {
						"required": true,
						"content": { "application/json": { "schema": { "$ref": "#/components/schemas/KontaktInput" } } }
					},
					"responses": { "200": { "description": "E-post sendt" }, "400": { "description": "Valideringsfeil" }, "500": { "description": "E-postfeil" } }
				}
			},
			"/byer": { "get": { "summary": "Hent byer", "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object", "additionalProperties": true } } } } }, "500": { "description": "Serverfeil" } } } },
			"/BetaleAbo": { "post": { "summary": "Marker abonnement som betalt", "security": [{ "cookieAuth": [] }], "responses": { "200": { "description": "Registrert" }, "404": { "description": "Ingen endringer" }, "500": { "description": "Serverfeil" } } } },
			"/AvslutteAbo": { "post": { "summary": "Avslutt abonnement", "security": [{ "cookieAuth": [] }], "responses": { "200": { "description": "Avsluttet" }, "404": { "description": "Ingen endringer" }, "500": { "description": "Serverfeil" } } } },
			"/tommeTestdata": { "delete": { "summary": "Tøm testdata (Klubber og Reviews)", "responses": { "200": { "description": "Tømt" }, "500": { "description": "Serverfeil" } } } },
...
	},
	"components": {
		"securitySchemes": {
			"cookieAuth": { "type": "apiKey", "in": "cookie", "name": "connect.sid" }
		},
		"schemas": {
...
			"Systemlogg": { "type": "object", "properties": { "_id": { "type": "string" }, "tidspunkt": { "type": "string", "format": "date-time" }, "bruker": { "type": "string" }, "handling": { "type": "string" }, "detaljer": { "type": "string" } } },
			"SystemloggInput": { "type": "object", "required": ["bruker", "handling"], "properties": { "bruker": { "type": "string" }, "handling": { "type": "string" }, "detaljer": { "type": "string" } } },
...
			"SessionBruker": { "type": "object", "properties": { "bruker": { "$ref": "#/components/schemas/BrukerDetalj" } } },
			"BrukerDetalj": { "type": "object", "properties": { "id": { "type": "string" }, "brukernavn": { "type": "string" }, "epost": { "type": "string" }, "fornavn": { "type": "string" }, "etternavn": { "type": "string" }, "telefonnummer": { "type": "string" }, "bosted": { "type": "string" }, "rolle": { "type": "string" }, "poengkort": { "type": "array", "items": { "$ref": "#/components/schemas/Poengkort" } }, "invitasjoner": { "type": "array", "items": { "$ref": "#/components/schemas/Invitasjon" } }, "betalt": { "type": "boolean" } } },
			"KontaktInput": { "type": "object", "required": ["navn", "epost", "melding"], "properties": { "navn": { "type": "string" }, "epost": { "type": "string" }, "melding": { "type": "string" } } },
...
			"Poengkort": { "type": "object", "additionalProperties": true },
			"Invitasjon": { "type": "object", "additionalProperties": true },
...
		}
	}
}